Rearrange some code.
